refactor: move allow_button_presses to special2
