+++ src/main/java/ti4/commands/player/PlayerCommand.java
@@ -35,7 +35,6 @@ public class PlayerCommand implements ParentCommand {
             new AddTeamMate(),
             new RemoveTeamMate(),
             new SetStatsAnchor(),
-            new AllowButtonPresses(),
             new CCsButton()
     ).collect(Collectors.toMap(Subcommand::getName, subcommand -> subcommand));
