+++ src/main/java/ti4/commands/special/AllowButtonPresses.java
@@ -1,4 +1,4 @@
-package ti4.commands.player;
+package ti4.commands.special;
 
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
@@ -9,7 +9,7 @@ import ti4.helpers.Constants;
 import ti4.map.Player;
 import ti4.message.MessageHelper;
 
-class AllowButtonPresses extends GameStateSubcommand {
+public class AllowButtonPresses extends GameStateSubcommand {
 
     public AllowButtonPresses() {
         super(Constants.ALLOW_BUTTON_PRESSES, "Allow or disallow others to press buttons for a player", true, true);
